Create synapse table version
