+++ scripts/case_selection/update_case_count_table.py
@@ -8,7 +8,7 @@ import synapseclient
 from synapseclient import Table
 import yaml
 
-from utils import get_production, get_adjusted, get_pressure, get_sdv, get_irr
+from utils import get_production, get_adjusted, get_pressure, get_sdv, get_irr, create_synapse_table_version
 
 
 def get_current_production_record_count(
@@ -25,60 +25,6 @@ def get_current_production_record_count(
     record_ids = results.asDataFrame()
 
     return len(record_ids)
-
-
-def clear_synapse_table(syn, table_id):
-    """
-    Clears a synapse table by deleting all rows.
-
-    Parameters:
-        table_id (str): The ID of the table to be cleared.
-
-    Returns:
-        int: The number of rows deleted.
-    """
-    results = syn.tableQuery(f"SELECT * FROM {table_id}")
-    data = results.asDataFrame()
-    syn.delete(data)
-    return len(data)
-
-
-def update_synapse_table(syn, table_id, data):
-    """
-    Updates a table in the Synapse platform with new data.
-
-    Parameters:
-        table_id (str): The ID of the table to be updated.
-        data (pandas.DataFrame): The new data to be stored in the table.
-
-    Returns:
-        int: The number of rows in the updated table.
-    """
-    entity = syn.get(table_id)
-    project_id = entity.properties.parentId
-    table_name = entity.properties.name
-    cols = data.columns
-    table = Table(table_name, data, parent=project_id, columns=cols)
-    syn.store(table)
-    return len(data)
-
-
-def create_new_table_version(syn, table_id, data, comment=""):
-    """
-    Create a new version of a table in Synapse.
-
-    Parameters:
-        table_id (str): The ID of the table to create a new version for.
-        data (pandas.DataFrame): The data to update the table with.
-        comment (str, optional): A comment to associate with the new version.
-
-    Returns:
-        int: The version number of the newly created version.
-    """
-    _ = clear_synapse_table(syn, table_id)
-    _ = update_synapse_table(syn, table_id, data)
-    n_version = syn.create_snapshot_version(table_id, comment=comment)
-    return n_version
 
 
 def main():
@@ -171,7 +117,7 @@ def main():
 
     # save ------------------------
     if save_synapse:
-        n_version = create_new_table_version(
+        n_version = create_synapse_table_version(
             syn=syn,
             table_id=config["synapse"]["case_selection"]["id"],
             data=case_count_table_df,
